Reimplement get_cached_or_compile with futures
This commit just takes the synchronous version and reimplements it with
futures, using necessary combinators and such to prepare for running
commands with futures as well.

All process spawning/waiting still happens in a thread pool, this
commit just refactors methods to push futures all the way to down to the
lowest levels so we can integrate `tokio-process`.